move some code out of enc.cc
new files: dichotomy.cc (all the -psnr and -size search)
and headers.cc (all the WriteXXX() chunk-writing functions)
