+++ CMakeLists.txt
@@ -41,8 +41,10 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src/ ${SJPEG_DEP_INCLUDE_DIRS})
 add_library(sjpeg ${CMAKE_CURRENT_SOURCE_DIR}/src/bit_writer.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/bit_writer.h
   ${CMAKE_CURRENT_SOURCE_DIR}/src/colors_rgb.cc
-  ${CMAKE_CURRENT_SOURCE_DIR}/src/fdct.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/dichotomy.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/enc.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/fdct.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/headers.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/jpeg_tools.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/score_7.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/sjpeg.h
