+++ src/enc.cc
@@ -36,11 +36,6 @@ static const int32_t kDefaultBias = 0x78;
 static const int kDefaultDeltaMaxLuma = 12;
 static const int kDefaultDeltaMaxChroma = 1;
 
-// convergence is considered reached if |dq| < kdQLimit
-static const float kdQLimit = 0.2;
-// maximal variation allowed on dQ
-static const float kdQThresh = 200.;
-
 // finer tuning of perceptual optimizations:
 
 // Minimum average number of entries per bin required for performing histogram-
@@ -113,8 +108,7 @@ void SetQuantMatrix(const uint8_t in[64], int q_factor, uint8_t out[64]) {
   }
 }
 
-void SetMinQuantMatrix(const uint8_t* const m, uint8_t out[64],
-                       int tolerance) {
+void SetMinQuantMatrix(const uint8_t* const m, uint8_t out[64], int tolerance) {
   assert(out != NULL);
   if (m != NULL) {
     for (int i = 0; i < 64; ++i) {
@@ -127,6 +121,41 @@ void SetMinQuantMatrix(const uint8_t* const m, uint8_t out[64],
     }
   }
 }
+
+////////////////////////////////////////////////////////////////////////////////
+// Encoder main class
+
+Encoder::Encoder(int W, int H, int step, const uint8_t* const rgb)
+  : W_(W), H_(H), step_(step),
+    rgb_(rgb),
+    bw_(W_ * H_ / 4),      // very reasonable guess about final size
+    in_blocks_base_(NULL),
+    in_blocks_(NULL),
+    have_coeffs_(false),
+    passes_(1),
+    min_psnr_(0.),
+    all_run_levels_(NULL),
+    nb_run_levels_(0),
+    max_run_levels_(0),
+    qdelta_max_luma_(kDefaultDeltaMaxLuma),
+    qdelta_max_chroma_(kDefaultDeltaMaxChroma) {
+  SetCompressionMethod(kDefaultMethod);
+  SetQuality(kDefaultQuality);
+  SetYUVFormat(false);
+  SetQuantizationBias(kDefaultBias, false);
+  const uint8_t* tmp[2] = { NULL, NULL };
+  SetMinQuantMatrices(tmp, 0);
+  InitializeStaticPointers();
+  memset(dc_codes_, 0, sizeof(dc_codes_));  // safety
+  memset(ac_codes_, 0, sizeof(ac_codes_));
+}
+
+Encoder::~Encoder() {
+  delete[] all_run_levels_;
+  DesallocateBlocks();   // clean-up leftovers in case of we had an error
+}
+
+////////////////////////////////////////////////////////////////////////////////
 
 void Encoder::SetQuality(int q) {
   q = GetQFactor(q);
@@ -211,38 +240,6 @@ void Encoder::InitializeStaticPointers() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-
-Encoder::Encoder(int W, int H, int step, const uint8_t* const rgb)
-  : W_(W), H_(H), step_(step),
-    rgb_(rgb),
-    bw_(W_ * H_ / 4),      // very reasonable guess about final size
-    in_blocks_base_(NULL),
-    in_blocks_(NULL),
-    have_coeffs_(false),
-    passes_(1),
-    min_psnr_(0.),
-    all_run_levels_(NULL),
-    nb_run_levels_(0),
-    max_run_levels_(0),
-    qdelta_max_luma_(kDefaultDeltaMaxLuma),
-    qdelta_max_chroma_(kDefaultDeltaMaxChroma) {
-  SetCompressionMethod(kDefaultMethod);
-  SetQuality(kDefaultQuality);
-  SetYUVFormat(false);
-  SetQuantizationBias(kDefaultBias, false);
-  const uint8_t* tmp[2] = { NULL, NULL };
-  SetMinQuantMatrices(tmp, 0);
-  InitializeStaticPointers();
-  memset(dc_codes_, 0, sizeof(dc_codes_));  // safety
-  memset(ac_codes_, 0, sizeof(ac_codes_));
-}
-
-Encoder::~Encoder() {
-  delete[] all_run_levels_;
-  DesallocateBlocks();   // clean-up leftovers in case of we had an error
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // memory and internal buffers managment. We grow on demand.
 
 void Encoder::CheckBuffers() {
@@ -282,96 +279,6 @@ void Encoder::DesallocateBlocks() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Headers
-//
-// NOTE(skal): all chunks start with a startcode '0xff??' (0xffd8 e.g),
-// followed by the size of the payload *not counting the startcode*!
-// That's why you often find these 'Reserve(data_size + 2)' below, the '+2'
-// accounting for the 0xff?? startcode size.
-
-void Encoder::WriteAPP0() {  // SOI + APP0
-  const uint8_t kHeader0[] = {
-    0xff, 0xd8,                     // SOI
-    0xff, 0xe0, 0x00, 0x10,         // APP0
-    0x4a, 0x46, 0x49, 0x46, 0x00,   // 'JFIF'
-    0x01, 0x01,                     // v1.01
-    0x00, 0x00, 0x01, 0x00, 0x01,   // aspect ratio = 1:1
-    0x00, 0x00                      // thumbnail width/height
-  };
-  bw_.Reserve(sizeof(kHeader0));
-  bw_.PutBytes(kHeader0, sizeof(kHeader0));
-}
-
-bool Encoder::WriteAPPMarkers(const std::string& data) {
-  if (data.size() == 0) return true;
-  const size_t data_size = data.size();
-  bw_.Reserve(data_size);
-  bw_.PutBytes(reinterpret_cast<const uint8_t*>(data.data()), data.size());
-  return true;
-}
-
-bool Encoder::WriteEXIF(const std::string& data) {
-  if (data.size() == 0) return true;
-  const uint8_t kEXIF[] = "Exif\0";
-  const size_t kEXIF_len = 6;  // includes the \0's
-  const size_t data_size = data.size() + kEXIF_len + 2;
-  if (data_size > 0xffff) return false;
-  bw_.Reserve(data_size);
-  bw_.PutByte(0xff);
-  bw_.PutByte(0xe1);
-  bw_.PutByte((data_size >> 8) & 0xff);
-  bw_.PutByte((data_size >> 0) & 0xff);
-  bw_.PutBytes(kEXIF, kEXIF_len);
-  bw_.PutBytes(reinterpret_cast<const uint8_t*>(data.data()), data.size());
-  return true;
-}
-
-bool Encoder::WriteICCP(const std::string& data) {
-  if (data.size() == 0) return true;
-  size_t data_size = data.size();
-  const uint8_t* ptr = reinterpret_cast<const uint8_t*>(data.data());
-  const uint8_t kICCP[] = "ICC_PROFILE";
-  const size_t kICCP_len = 12;  // includes the \0
-  const size_t chunk_size_max = 0xffff - kICCP_len - 4;
-  size_t max_chunk = (data_size + chunk_size_max - 1) / chunk_size_max;
-  if (max_chunk >= 256) return false;
-  size_t seq = 1;
-  while (data_size > 0) {
-    size_t size = data_size;
-    if (size > chunk_size_max) size = chunk_size_max;
-    bw_.Reserve(size + kICCP_len + 4 + 2);
-    bw_.PutByte(0xff);
-    bw_.PutByte(0xe2);
-    bw_.PutByte(((size + kICCP_len + 4) >> 8) & 0xff);
-    bw_.PutByte(((size + kICCP_len + 4) >> 0) & 0xff);
-    bw_.PutBytes(kICCP, kICCP_len);
-    bw_.PutByte(seq & 0xff);
-    bw_.PutByte(max_chunk & 0xff);
-    bw_.PutBytes(ptr, size);
-    ptr += size;
-    data_size -= size;
-    seq += 1;
-  }
-  return true;
-}
-
-bool Encoder::WriteXMP(const std::string& data) {
-  if (data.size() == 0) return true;
-  const uint8_t kXMP[] = "http://ns.adobe.com/xap/1.0/";
-  const size_t kXMP_size = 29;
-  const size_t data_size = 2 + data.size() + kXMP_size;
-  if (data_size > 0xffff) return false;  // error
-  bw_.Reserve(data_size + 2);
-  bw_.PutByte(0xff);
-  bw_.PutByte(0xe1);
-  bw_.PutByte((data_size >> 8) & 0xff);
-  bw_.PutByte((data_size >> 0) & 0xff);
-  bw_.PutBytes(kXMP, kXMP_size);
-  bw_.PutBytes(reinterpret_cast<const uint8_t*>(data.data()), data.size());
-  return true;
-}
-
-////////////////////////////////////////////////////////////////////////////////
 
 #define FP_BITS 16    // fractional precision for fixed-point dividors
 #define AC_BITS 4     // extra precision bits from fdct's scaling
@@ -407,42 +314,6 @@ void Encoder::FinalizeQuantMatrix(Quantizer* const q, int q_bias) {
 
 void Encoder::SetCostCodes(int idx) {
   quants_[idx].codes_ = ac_codes_[idx];
-}
-
-void Encoder::WriteDQT() {
-  const int data_size = 2 * 65 + 2;
-  const uint8_t kDQTHeader[] = { 0xff, 0xdb, 0x00, (uint8_t)data_size };
-  bw_.Reserve(data_size + 2);
-  bw_.PutBytes(kDQTHeader, sizeof(kDQTHeader));
-  for (int n = 0; n <= 1; ++n) {
-    bw_.PutByte(n);
-    const uint8_t* quant = quants_[n].quant_;
-    for (int i = 0; i < 64; ++i) {
-      bw_.PutByte(quant[kZigzag[i]]);
-    }
-  }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-#define DATA_16b(X) ((uint8_t)((X) >> 8)), ((uint8_t)((X) & 0xff))
-
-void Encoder::WriteSOF() {   // SOF
-  const int data_size = 8 + 3 * nb_comps_;
-  assert(data_size <= 255);
-  const uint8_t kHeader[] = {
-    0xff, 0xc0, DATA_16b(data_size),         // SOF0 marker, size
-    0x08,                                    // 8bits/components
-    DATA_16b(H_), DATA_16b(W_),              // height, width
-    (uint8_t)nb_comps_                       // number of components
-  };
-  bw_.Reserve(data_size + 2);
-  bw_.PutBytes(kHeader, sizeof(kHeader));
-  for (int c = 0; c < nb_comps_; ++c) {
-    bw_.PutByte(c + 1);
-    bw_.PutByte(block_dims_[c]);
-    bw_.PutByte(quant_idx_[c]);
-  }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -541,55 +412,6 @@ void Encoder::InitCodes(bool only_ac) {
       (void)nb_syms;
     }
   }
-}
-
-void Encoder::WriteDHT() {
-  InitCodes(false);
-  const int nb_tables = (nb_comps_ == 1 ? 1 : 2);
-  for (int c = 0; c < nb_tables; ++c) {   // luma, chroma
-    for (int type = 0; type <= 1; ++type) {               // dc, ac
-      const HuffmanTable* const h = Huffman_tables_[type * 2 + c];
-      const int data_size = 3 + 16 + h->nb_syms_;
-      assert(data_size <= 255);
-      bw_.Reserve(data_size + 2);
-      bw_.PutByte(0xff);
-      bw_.PutByte(0xc4);
-      bw_.PutByte(0x00 /*data_size >> 8*/);
-      bw_.PutByte(data_size);
-      bw_.PutByte((type << 4) | c);
-      bw_.PutBytes(h->bits_, 16);
-      bw_.PutBytes(h->syms_, h->nb_syms_);
-    }
-  }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-void Encoder::WriteSOS() {   // SOS
-  const int data_size = 6 + nb_comps_ * 2;
-  assert(data_size <= 255);
-  const uint8_t kHeader[] = {
-      0xff, 0xda, DATA_16b(data_size), (uint8_t)nb_comps_
-  };
-  bw_.Reserve(data_size + 2);
-  bw_.PutBytes(kHeader, sizeof(kHeader));
-  for (int c = 0; c < nb_comps_; ++c) {
-    bw_.PutByte(c + 1);
-    bw_.PutByte(quant_idx_[c] * 0x11);
-  }
-  bw_.PutByte(0x00);        // Ss
-  bw_.PutByte(0x3f);        // Se
-  bw_.PutByte(0x00);        // Ah/Al
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-void Encoder::WriteEOI() {   // EOI
-  bw_.Flush();
-  // append EOI
-  bw_.Reserve(2);
-  bw_.PutByte(0xff);
-  bw_.PutByte(0xd9);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1302,7 +1124,7 @@ void Encoder::FinalPassScan(size_t nb_mbs, const DCTCoeffs* coeffs) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Multi-pass
+// Huffman tables optimization
 
 void Encoder::ResetEntropyStats() {
   memset(freq_ac_, 0, sizeof(freq_ac_));
@@ -1600,280 +1422,6 @@ void Encoder::SinglePassScanOptimized() {
     FinalPassScan(nb_mbs, base_coeffs);
   }
   delete[] base_coeffs;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// dichotomy
-
-#define DBG_PRINT 0
-
-struct PassStats {  // struct for organizing convergence in either size or PSNR
-  bool is_first;
-  float dq, q, last_q;
-  double value, last_value, target;
-  bool do_size_search;
-
-  PassStats(const Encoder& enc);
-  bool ComputeNextQ(float result);   // returns true if finished
-  void BackTrack() {
-    q = last_q;
-    dq /= 2;
-    q += dq;
-  }
-};
-
-PassStats::PassStats(const Encoder& enc) {
-  is_first = true;
-  dq = 50.f;
-  q = last_q = 100.;
-  value = last_value = 0;
-  target = enc.target_value_;
-  do_size_search = (enc.target_mode_ == SjpegEncodeParam::TARGET_SIZE);
-}
-
-static float Clamp(float v, float min, float max) {
-  return (v < min) ? min : (v > max) ? max : v;
-}
-
-bool PassStats::ComputeNextQ(float result) {
-  value = result;
-  if (is_first) {
-    if (do_size_search) {
-      dq = (value < target) ? -dq : dq;
-    } else {
-      dq = (value < target) ? -dq : dq;
-    }
-    is_first = false;
-  } else {
-    if (fabs(value - last_value) > 0.01 * value) {
-      const double slope = (target - value) / (last_value - value);
-      dq = (float)(slope * (last_q - q));
-    } else {
-      dq = 0.;  // we're done?!
-    }
-  }
-  // Limit variable to avoid large swings.
-  dq = Clamp(dq, -kdQThresh, kdQThresh);
-  last_q = q;
-  last_value = value;
-  q = Clamp(q + dq, 0.f, 2000.f);
-  return fabs(q - last_q) < kdQLimit;
-}
-
-void Encoder::LoopScan() {
-  assert(use_extra_memory_);
-  assert(reuse_run_levels_);
-
-  if (use_adaptive_quant_) {
-    CollectHistograms();
-  } else {
-    CollectCoeffs();   // we just need the coeffs
-  }
-
-  const QuantizeBlockFunc quantize_block = use_trellis_ ? TrellisQuantizeBlock
-                                                        : quantize_block_;
-  // We use the default Huffman tables as basis for bit-rate evaluation
-  if (use_trellis_) InitCodes(true);
-
-  const size_t nb_mbs = mb_w_ * mb_h_ * mcu_blocks_;
-  DCTCoeffs* const base_coeffs = new DCTCoeffs[nb_mbs];
-  uint8_t base_quant[2][64], opt_quants[2][64];
-  for (int c = 0; c < 2; ++c) {
-    CopyQuantMatrix(quants_[c].quant_, base_quant[c]);
-  }
-
-  // Dichotomy passes
-  PassStats stats(*this);
-  for (int p = 0; p < passes_; ++p) {
-    for (int c = 0; c < 2; ++c) {
-      SetQuantMatrix(base_quant[c], stats.q, quants_[c].quant_);
-      FinalizeQuantMatrix(&quants_[c], q_bias_);
-    }
-
-    if (use_adaptive_quant_) {
-      AnalyseHisto();   // adjust quant_[] matrices
-    }
-
-    // compute pass to store coeffs / runs / dc_code_
-    ResetDCs();
-    nb_run_levels_ = 0;
-    int16_t* in = in_blocks_;
-    DCTCoeffs* coeffs = base_coeffs;
-    for (int n = 0; n < mb_w_ * mb_h_; ++n) {
-      CheckBuffers();
-      for (int c = 0; c < nb_comps_; ++c) {
-        for (int i = 0; i < nb_blocks_[c]; ++i) {
-          RunLevel* const run_levels = all_run_levels_ + nb_run_levels_;
-          const int dc = quantize_block(in, c, &quants_[quant_idx_[c]],
-                                        coeffs, run_levels);
-          coeffs->dc_code_ = GenerateDCDiffCode(dc, &DCs_[c]);
-          nb_run_levels_ += coeffs->nb_coeffs_;
-          ++coeffs;
-          in += 64;
-        }
-      }
-    }
-    if (optimize_size_) {
-      StoreOptimalHuffmanTables(nb_mbs, base_coeffs);
-      if (use_trellis_) InitCodes(true);
-    }
-
-    const float result =
-      stats.do_size_search ? ComputeSize(base_coeffs, all_run_levels_)
-                           : ComputePSNR(base_coeffs, all_run_levels_);
-    if (p > 0 && min_psnr_ > 0.) {
-      const float psnr =
-          stats.do_size_search ? ComputePSNR(base_coeffs, all_run_levels_)
-                               : result;
-      if (psnr < min_psnr_) {
-        stats.BackTrack();
-        continue;
-      }
-    }
-    if (DBG_PRINT) printf("pass #%d: q=%f value:%.2f\n", p, stats.q, result);
-    for (int c = 0; c < 2; ++c) {
-      CopyQuantMatrix(quants_[c].quant_, opt_quants[c]);
-    }
-    if (stats.ComputeNextQ(result)) break;
-  }
-  SetQuantMatrices(opt_quants);   // set the final matrix
-
-  // finish bitstream
-  WriteDQT();
-  WriteSOF();
-  WriteDHT();
-  WriteSOS();
-  FinalPassScan(nb_mbs, base_coeffs);
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Size & PSNR computation, mostly for dichotomy
-
-size_t Encoder::HeaderSize() const {
-  size_t size = 0;
-  size += 20;    // APP0
-  size += app_markers_.size();
-  if (exif_.size() > 0) {
-    size += 8 + exif_.size();
-  }
-  if (iccp_.size() > 0) {
-    const size_t chunk_size_max = 0xffff - 12 - 4;
-    const size_t num_chunks = (iccp_.size() - 1) / chunk_size_max + 1;
-    size += num_chunks * (12 + 4 + 2);
-    size += iccp_.size();
-  }
-  if (xmp_.size() > 0) {
-    size += 2 + 2 + 29 + xmp_.size();
-  }
-  size += 2 * 65 + 2 + 2;         // DQT
-  size += 8 + 3 * nb_comps_ + 2;  // SOF
-  size += 6 + 2 * nb_comps_ + 2;  // SOS
-  size += 2;                      // EOI
-  // DHT:
-  for (int c = 0; c < (nb_comps_ == 1 ? 1 : 2); ++c) {   // luma, chroma
-    for (int type = 0; type <= 1; ++type) {               // dc, ac
-      const HuffmanTable* const h = Huffman_tables_[type * 2 + c];
-      size += 2 + 3 + 16 + h->nb_syms_;
-    }
-  }
-  return size * 8;
-}
-
-void Encoder::BlocksSize(int nb_mbs, const DCTCoeffs* coeffs,
-                              const RunLevel* rl,
-                              BitCounter* const bc) const {
-  for (int n = 0; n < nb_mbs; ++n) {
-    const DCTCoeffs& c = coeffs[n];
-    const int idx = c.idx_;
-    const int q_idx = quant_idx_[idx];
-
-    // DC
-    const int dc_len = c.dc_code_ & 0x0f;
-    const uint32_t code = dc_codes_[q_idx][dc_len];
-    bc->AddPackedCode(code);
-    bc->AddBits(c.dc_code_ >> 4, dc_len);
-
-    // AC
-    const uint32_t* const codes = ac_codes_[q_idx];
-    for (int i = 0; i < c.nb_coeffs_; ++i) {
-      int run = rl[i].run_;
-      while (run & ~15) {        // escapes
-        bc->AddPackedCode(codes[0xf0]);
-        run -= 16;
-      }
-      const uint32_t suffix = rl[i].level_;
-      const size_t nbits = suffix & 0x0f;
-      const int sym = (run << 4) | nbits;
-      bc->AddPackedCode(codes[sym]);
-      bc->AddBits(suffix >> 4, nbits);
-    }
-    if (c.last_ < 63) bc->AddPackedCode(codes[0x00]);  // EOB
-    rl += c.nb_coeffs_;
-  }
-}
-
-float Encoder::ComputeSize(const DCTCoeffs* coeffs,
-                                 const RunLevel* rl) {
-  InitCodes(false);
-  size_t size = HeaderSize();
-  BitCounter bc;
-  BlocksSize(mb_w_ * mb_h_ * mcu_blocks_, coeffs, rl, &bc);
-  size += bc.Size();
-  return size / 8.f;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-static double GetPSNR(uint64_t err, uint64_t size) {
-  return (err > 0 && size > 0) ? 10. * log10(255. * 255. * size / err) : 99.;
-}
-
-static const uint16_t kMasks[16 + 1][2] = {
-  { 0x0000U, 0x0000U },  // fake extra entry, for robustness
-  { 0x0001U, 0x0001U }, { 0x0002U, 0x0003U },
-  { 0x0004U, 0x0007U }, { 0x0008U, 0x000fU },
-  { 0x0010U, 0x001fU }, { 0x0020U, 0x003fU },
-  { 0x0040U, 0x007fU }, { 0x0080U, 0x00ffU },
-  { 0x0100U, 0x01ffU }, { 0x0200U, 0x03ffU },
-  { 0x0400U, 0x07ffU }, { 0x0800U, 0x03ffU },
-  { 0x1000U, 0x1fffU }, { 0x2000U, 0x3fffU },
-  { 0x4000U, 0x7fffU }, { 0x8000U, 0xffffU }
-};
-
-static int16_t CodeToCoeff(uint16_t code) {
-  const int nbits = (code & 0xf);
-  int16_t qcoeff = (code >> 4);
-  if (qcoeff < kMasks[nbits][0]) qcoeff -= kMasks[nbits][1];
-  return qcoeff;
-}
-
-float Encoder::ComputePSNR(const DCTCoeffs* coeffs,
-                                const RunLevel* rl) const {
-  uint64_t error = 0;
-  int16_t* in = in_blocks_;
-  const size_t nb_mbs = mb_w_ * mb_h_ * mcu_blocks_;
-  int32_t DCs[3] = { 0, 0, 0 };
-  for (size_t n = 0; n < nb_mbs; ++n) {
-    const DCTCoeffs& c = coeffs[n];
-    const int c_idx = c.idx_;
-    const int q_idx = quant_idx_[c_idx];
-    const uint8_t* const Q = quants_[q_idx].quant_;
-    int pos = 0;
-    int32_t dq[64] = { 0 };
-    for (int i = 0; i < c.nb_coeffs_; ++i, ++rl) {
-      pos += rl->run_ + 1;
-      const int j = kZigzag[pos];
-      dq[j] = Q[j] * CodeToCoeff(rl->level_);
-    }
-    dq[0] = (DCs[c_idx] += CodeToCoeff(c.dc_code_)) * Q[0];
-    for (int i = 0; i < 64; ++i) {
-      const int16_t v0 = in[i] / 16;  // fDCT output is upscaled
-      const int16_t v = dq[i];
-      error += ((uint64_t)(v0 - v) * (v0 - v));
-    }
-    in += 64;
-  }
-  return GetPSNR(error, nb_mbs * 64ull);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
